Create parent directory of output when necessary
